requirements.txt
asgiref==3.3.1
Django==3.1.5
pytz==2020.5
sqlparse==0.4.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.1
python-dotenv==0.15.0
coreapi==2.3.3
coreschema==0.0.4
certifi==2020.12.5
chardet==4.0.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
requests==2.25.1
uritemplate==3.0.1
urllib3==1.26.2
psycopg2-binary==2.8.6
django-cors-headers==3.5.0
drf-yasg==1.20.0
inflection==0.5.1
packaging==20.8
pyparsing==2.4.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
